Allow a single connection from 127.0.0.1 as long as it isn't from self.
We still need to support multiple connections, as detailed by 
https://github.com/MerosCrypto/Meros/issues/159, yet that requires more 
work for a few reasons.

The main problem is we can't link a live socket with a sync socket 
anymore. This creates the idea of multiple peers when there's really 
only one. Without adding an ID to the handshake, all we can do is add 
the new socket to the latest localhost peer who's missing that socket. 
It's not impossible, just imperfect.