Add a FileLimitTracker.
Keeps track of the amount of file Meros is allowed to use and how many 
it is using. Uses a mix of regular updating and estimation of how many 
new files have been created.

When Meros is `BUFFER_FILES` near its limit, it starts recommending not 
handling peers/not attempting connections.

Should work on Linux and macOS. Windows is not currently supported, 
though the WINAPI functions seem to be well documented. Adds a 
compile-time UnsupportedTarget error which is raised when compilation is 
attempted on an unsupport target (mainly Windows, but technically 
Solaris and Haiku as well).