ULimit (#169)
* Add a FileLimitTracker.

Keeps track of the amount of file Meros is allowed to use and how many 
it is using. Uses a mix of regular updating and estimation of how many 
new files have been created.

When Meros is `BUFFER_FILES` near its limit, it starts recommending not 
handling peers/not attempting connections.

Should work on Linux and macOS. Windows is not currently supported, 
though the WINAPI functions seem to be well documented. Adds a 
compile-time UnsupportedTarget error which is raised when compilation is 
attempted on an unsupport target (mainly Windows, but technically 
Solaris and Haiku as well).

* Don't create connections when we don't have files available.

Also sends Busy to people who connect to us when we don't have files 
available, a currently undocumented network message equivalent to Peers 
yet only used as a HandshakeSyncing response.

* Handle Busy messages.

* Bug fixes.

* Update the documentation to include Busy.

Also fixes some misc errors in the docs.

* Add a test to verify Meros sends Busy/manages FDs properly.

This test doesn't verify FDs are never leaked, which is exceptionally 
difficult to do. It creates sockets until it gets Busy, creates 2x the 
buffer to make sure busy sockets are closed, disconnects 50 sockets, 
reconnects them, and verifies any new socket gets Busy.

I originally thought this test discovered a FD leak in the form of the 
timers. 3 timers are set during the test, and Meros could only reconnect 
47 sockets. That said, it's actually an issue with our original 
presumption of how many FDs are open. mainRPC is called AFTER the 
Network is connected and opens 3 new FDs.

This test now (compared to my first draft) waits a minute before running 
for Meros to correct itself. I added clearTimer statements in case they 
helped, and they are left because they cannot hurt.

I don't feel a need to fix the 3-file difference in Meros because I 
doubt 495 unique IPs will connect within the first minute of operation. 
Even if they did, we'd still have 13 buffer FDs.

* Remove a stray print.

* Make sure Meros properly handles Busy messages.

Also tweaks the ULimit test to make sure it passes the checks we employ.