Update to Nim 1.2.0.
I honestly regret this. Nim 1.0.0 didn't offer bug fixes or feature 
stability. Nim 1.2.0 continues with breaking existing code. While some 
of this is for the best, some of this isn't.

I initially tried to update to 1.0.6 with specifying old versions of 
Status packages as they've updated to require 1.2.0. That said, it ended 
up being too complicated, especially as we should update for performance 
and stability reasons.

All tests pass. New warnings occur due to the use of csize, which has 
been deprecated for 1.2.0. I did try to update to the new csize_t, which 
is a type I vehemently disagree with, yet csize_t is not compatible with 
csize. Until our packages are updated, we can't update Meros to csize_t.