Don't attempt naive wallet derivation.
We'd have to write a compliant Ed25519 BIP32 library or import one. The 
first isn't worth it, and I can't find one which already exists.

This does mean we can't check the key/signature of the Claim, but I'm 
not too concerned about that loss of functionality.

Also fixes a bug in checking if we were disconnected/lack of handling of 
the Claim Verification.