Fix #177. (#180)
* Prototype implementation.

Against the general procedure, this puts code before the protocol. This 
is done because I rather confirm viability of usage, with a good idea of 
the protocol, which is documented in #177, before formalizing it just to 
solely edit later.

* Move when we verify the Block-created Data; don't check their inputs.

* Prototype #177 test.

* Remove the blocking #178 from this branch by merging back the fix. (#179)

* Fix https://github.com/MerosCrypto/Meros/issues/178.

* Correct spacing for the HDWallet.

* Don't attempt naive wallet derivation.

We'd have to write a compliant Ed25519 BIP32 library or import one. The 
first isn't worth it, and I can't find one which already exists.

This does mean we can't check the key/signature of the Claim, but I'm 
not too concerned about that loss of functionality.

Also fixes a bug in checking if we were disconnected/lack of handling of 
the Claim Verification.

* Correct calls which fail due to API changes.

* Add to the protocol that each Block creates a Data upon its addition.

The specification of this Data, as well as its reasons, are included.

* Fix the 155 test.

* Fix disconnect handling in the #177 test.

* Remove the BIP utils stub.