Fix #181. (#196)
* Test Minisketch's creation against the Python Pinsketch impl.

This does NOT test decoding, as that requires writing a new function, 
and differs from what's discussed in 
https://github.com/MerosCrypto/Meros/issues/181 a bit.

* Insignificant commit explaining why there's no merge and defining decode.

* Tweaks so style checks pass.

Implements the FIELD_BYTES helper as well, so we don't have to 
repeatedly use FIELD_BITS // 8.

* Start of a decode.

This commit is horrible. It starts to implement decode, but then I 
literally walk away. This isn't worth my time when there's more pressing 
issues.

I do truly want a pure Python impl, and this commit establishes the 
usage of pyfinite, but I'm leaving this indefinitely due to its low 
priority.

* Add a Minisketch test.

* [WIP] A draft submission (#195)

* a variable renamed to be the same as in the C++ source

* working on it

* working on it

* preliminary untested code ready

* bug fixes

* bug fixes

* test does not pass

* final changes before draft submission

* Implement a replica of the Minisketch lib instead of Pinsketch's ref lib

This has a lot of mutability manipulations that need to be cleaned up. 
qrt is hardcoded for a single value which works with a very specific 
sketch I was testing; still working on a proper implementation.

* Implement a proper qrt algorithm.

* Implement inv; remove pyfinite dependency.

* Clean monic and gcd.

* Correct qrt/inv code.

* square, polyMod, gcd, and findRoots bug fixes.

* Fix the Pinsketch code.

Still needs a lot of cleanup.

* Clean up Pinsketch.

* Correct and clean the 181 test.

* Support sketches with no differences.

Also corrects a list mutability error, misplaced parens, and adds a sort 
call.

* Clean the qrt code.

* Lower the amount of iterations used in the 181 test.

Co-authored-by: Victor Porton <porton@narod.ru>