Test tweaks so they pass.
The Blockchain test does need these tweaks reverted, sometime in the 
future, yet the reversion may not make it onto this branch, for reasons 
detailed in https://github.com/MerosCrypto/Meros/issues/183.

The Consensus test tweaks are required, but generating a proper State 
may be beneficial compared to this faux, manually manipulated, State.