Remove extraneous operations from the State; correct SpamFilter.
SpamFilter used state.merit, as was needed for the existing (effectively 
pointless due to these changes) SpamFilter test. While this should be 
fine, theoretically, as any locked participants will be removed, it's 
not preferable. The reason this error occured is from malformed State 
manipulation in the test, which has been corrected.