Fix https://github.com/MerosCrypto/Meros/issues/228.
While working on this, I caused Meros to segfault trying to sync data. 
That's why Socket.send now has an isNil check. The surrounding code is 
designed so this case never happens, or at least I thought I designed 
and wrote it like that. That said, there's obviously some edges, and 
this is an easy, efficient fix, that offers a great amount of safety.