Remove Lock/Unlock/Gas from the protocol.
The only reason I wanted these at mainnet was for atomic swap 
functionality. Thanks to https://github.com/MerosCrypto/asmr, this 
functionality is already present.

This functionality would've taken weeks to implement, not to mention the 
variety of bugs/testings that would need to be worked on.

Their message/prefix bytes are preserved so a MIP can add back this 
functionality in the future, after launching the mainnet.