Fix family discovery.
It was buggy in relation to Datas.

Also restores using calculateMerit when defaulting, which also verifies 
descendants when possible. This was removed for safety, despite 
theoretically being okay from the start. It was added back, in a safe 
way, as the ConsensusRevert tests relies on this behavior. While it is 
possible to simply supplement the CR test, this adds functionality and 
helps maintain parity between the CR test and main.

Also adds some extra comments to better document the code.