Merge families, implementing defaulting. (#251)
* Initial Family Management code

* Clarify how Epochs brings up Transactions

* Re-enable PruneUnaddable and add a family test

* Families fixes so PruneUnaddable works.

* Working family finalization.

Includes defaulting. UnmentionBeatMentioned works if the force failure 
at the end of it is removed.

Needs much more thorough testing before merging.

This, and the Epochs redo, points a lot of optimizations that can be 
made to this pipeline.

* Fix family equality checking.

Expands the Transactions test with the new calls required for the 
families to be properly updated.

* Remove inlined vector generation for UnmentionedBeatMentioned

* Add empty test files.

* Unioned families tests.

Includes a bug fix/some minor logging updates.

* Add a test meant for the last commit.

* Add a lower hash tie break test.

* Test more complex family behavior.

In this case, it's attempted verification of a child before its parent, 
while they're in the same family.

Replaces the old empty test files with new ones, better suited, 
including descriptions.

* Increase safety in family finalization.

Does add a potential slow down (children of transactions with 
competitors), yet the potential bug was multi-faceted and would require 
testing all of them to cover potential changes. Even then...

* Alternate version of previously added test.

* Add a test for an impossible transaction winning a family.

* Fix family discovery.

It was buggy in relation to Datas.

Also restores using calculateMerit when defaulting, which also verifies 
descendants when possible. This was removed for safety, despite 
theoretically being okay from the start. It was added back, in a safe 
way, as the ConsensusRevert tests relies on this behavior. While it is 
possible to simply supplement the CR test, this adds functionality and 
helps maintain parity between the CR test and main.

Also adds some extra comments to better document the code.