Massive rework. Will not compile on purpose.
Moves the Consensus and State to only care about a single Merit Removal.

Drops MeritRemoval entirely for just SignedMeritRemoval, which is no 
longer a descendant of Element.

Adds a `removals` field to Blocks defined as set[uint16].

This does remove the functionality where a Merit Removal can exist of a 
Transaction not addable to the current DAG. It's necessary, just as 
dropping VP MRs is.