Implicit Merit Removals (#258)
* Abstract checking if a Verification is malicious.

* Raise if a BlockBody contains a Merit Removal.

* Remove MeritRemoval VerificationPackets.

* Create a Merit Removal out of Same Nonce Merit Removals.

Doesn't support:

- Competing Verification Merit Removals

- Multiple Merit Removals. If A, B, and C share a nonce, then A/B, B/C, 
and A/C are valid Merit Removals. This matters for when we save the 
reasons, so the MR can't be replayed thanks to B/C if we only generated 
the MR for A/B.

* Work on generating Competing Verification Merit Removals.

* Massive rework. Will not compile on purpose.

Moves the Consensus and State to only care about a single Merit Removal.

Drops MeritRemoval entirely for just SignedMeritRemoval, which is no 
longer a descendant of Element.

Adds a `removals` field to Blocks defined as set[uint16].

This does remove the functionality where a Merit Removal can exist of a 
Transaction not addable to the current DAG. It's necessary, just as 
dropping VP MRs is.

* Cleanup and bug fixes.

* Further cleanup/bug fixes.

Notably with how Merit Holders are marked as malicious.

* Move former-Element equality checks.

* Update Nim tests.

Fixes equality checking errors due to Nim intracacies.

Doesn't update DB/Consensus/* tests.

* Nim tweaks.

Notably makes BlockBody a ref to stop the aggregate signature from being 
dropped while stored in a future.

* Update the Python tests.

Only regens + verifies the SameNonce test. Drops several tests which are 
no longer applicable.

* Update the protocol docs.

* Include implicit Merit Removals in the RPC's getBlock method.

* Regen JSON vectors.

* Updating typing on Nim tests.

* Fix when Merit Removals from Blocks are added to the Consensus DAG.

* Fix creation of VerifyCompeting MeritRemovals.

If the signed Verification was already tracked, it was swallowed.

* Remove the Invalid Competing test.

Now that all TXs used in a MR must be on-DAG, it's not relevant.

* Rebroadcast the new MR, not the first.

* Update Beaten to a behavioral change made.

* Fix Block templates involving partial MeritRemovals.

* Fix reloading the malicious table (and the accompanying test).

* Finish updating the Consensus test.

* Update the ConsensusRevert code.

This calls remove, yet it never generates any removals. The deletions in 
this diff alone should be enough. That said, parity has been maintained.