Initial Nim impl.
Removes significant from the header. Instead of including the packets 
quantity there, a SketchyBlockHeader object was introduced. This was 
because it'd be duplicating block.body.packets.len, and therefore 
requiring tracking twice. It's also only used when serializing. That 
said, since we serialize to hash, this may have been a poor decision.

Removes the Sketcher type for just seq[VerificationPacket] as we no 
longer track significance.