Update Python to the new BBR.
Helper method added for dealing with it. Old method to send a BlockBody 
renamed to rawBlockBody, which is used by the Syncer's loop and a single 
test which mirrors it. I did not put the work in to move both over.

204 has been radically changed. Its vectors expect to send a block and 
then multiple failure cases. That said, it only ever sends a single one. 
The test has been adjusted accordingly, and the vector has a comment 
added. Further cleanup would be beneficial.

240, added due to a bug in which we handle significant which has now 
been removed, has bene ported and remains. The top of the test now has a 
comment saying why.

Also corrects documentation in the PruneUnaddable test. It said the 
Verification was for the beaten descendant's child. The generator 
created one for the winning descendant though. The test's comment has 
been updated, rather than changing the vectors. What the original intent 
was, I do not remember.