Fix https://github.com/MerosCrypto/Meros/issues/274. (#277)
* Update protocol docs according to https://github.com/MerosCrypto/Meros/issues/274.

* Initial Nim impl.

Removes significant from the header. Instead of including the packets 
quantity there, a SketchyBlockHeader object was introduced. This was 
because it'd be duplicating block.body.packets.len, and therefore 
requiring tracking twice. It's also only used when serializing. That 
said, since we serialize to hash, this may have been a poor decision.

Removes the Sketcher type for just seq[VerificationPacket] as we no 
longer track significance.

* Discard SketchyBlockHeader for a BlockHeader with packetsQuantity.

See last commit description for more info.

* Update the tests.

Removes the DistinctUInt32 temporaily used to ensure lack of promotion 
for values representing significant in missed calls.

* Update the RPC docs.

* Update the Python.

VerifyCompeting now fails as it can successfully resolve the sketch. 
Considering sketches themselves weren't changed, this was probably a bug 
with significant no longer applying? Not sure. Still under review.

* Fix incorrect flow assumptions of VerifyCompeting.

* Ensure a valid packetsQuantity was set.

Adds matching tests.

* Add Nim support for the new BlockBodyRequest.

* Update Python to the new BBR.

Helper method added for dealing with it. Old method to send a BlockBody 
renamed to rawBlockBody, which is used by the Syncer's loop and a single 
test which mirrors it. I did not put the work in to move both over.

204 has been radically changed. Its vectors expect to send a block and 
then multiple failure cases. That said, it only ever sends a single one. 
The test has been adjusted accordingly, and the vector has a comment 
added. Further cleanup would be beneficial.

240, added due to a bug in which we handle significant which has now 
been removed, has bene ported and remains. The top of the test now has a 
comment saying why.

Also corrects documentation in the PruneUnaddable test. It said the 
Verification was for the beaten descendant's child. The generator 
created one for the winning descendant though. The test's comment has 
been updated, rather than changing the vectors. What the original intent 
was, I do not remember.

* Remove unused import.

* Add a test to verify Meros responds with the requested capacity.

* Update Block serialization tests.