Port the Transactions/Consensus/Merit/Network RPC modules.
Without the matching macro, this is mostly meaningless. Updates the docs 
accordingly.

Notably removed is the ability to broadcast DataDiffs/SendDiffs. This is 
due to the fact they're ordered and sequential, and at the same time, 
the protocol doesn't alloow requesting specific Elems. This means 
broadcasting Elem 5 will forever be doomed if Elem 4 didn't propagate.