Add requireAuth
Previously commented; now not in order to ensure it's not forgotten.

The GUI is always authed; the HTTP RPC's authed status is currently 
hardcoded to true.

Moves system_quit to its own module in order to enable access to 
requireAuth, a simulated pragma. While an auth check could've been used 
where it was originally, removing hardcoded methods is beneficial.

Correctly extends the RPC macro with requireAuth, and enables directly 
taking the request/reply function as arguments.

Adds a missing method not found error. Would previously only trigger if 
the module was found and the method wasn't.

Corrects the reply function to not supply the outer object, yet rather 
the macro generated reply call (as having it in the reply function 
itself broke errors).