Test getBlock.
Adds the packets field to the JSON, as Python needs it. While it was 
possible to remove it from Python, which was considered as this field 
has been previously excluded for being of little value, it is part of 
the header. If we ever add a getBlockHeader function, its value greatly 
increases.

Fixes an oversight on the range checks in retrieveFromJSON.