Correct Connection header parsing.
I assumed it'd just be keep-alive. It's actually a comma separated list 
which is generally just keep-alive.

Moves to a default policy of close. While that was the case for 1.0, 1.1 
defaults to keep-alive, when Meros expected close. While keep-alive will 
be supported, it's more efficient from a FD perspective to regain that 
slot ASAP.

Also utilizes a pair of logWarn statements for exceptions when shutting 
down.