Extend the seed test with getAddress (without arguments; unused) checks.
Uses the ed25519 RFC's signature code, ported to the existing ref impl's 
internal functions and cleaned up a bit (done due to an unknown issue 
with the original).

Slightly edits the private key management code of BLS.py in how it 
converts a 32-byte to a 48-byte value. Randomly complained it couldn't 
rjust with a byte char, despite working in the past and in my local 
python console.

Deletes the independent getAddress test for being irrelevant.