Test every personal method requires auth.
Previously, some methods weren't checked as they weren't implemented. 
Not that all are...

Requires moving the auth check to earlier in the function, which is a 
good thing in a couple of ways. Doesn't waste time arg parsing if not 
authed/less chance for something which changes data to happen.