Delete the samples folder.
The RPC Sample should be, and with this commit will be, deprecated for 
curl at this point.

The DB dump sample is irrelevant when we can just share a data folder 
and go from there. Sure, it's not in a JSON format, but it's easy enough 
to spin up a node and run the relevant commands as needed.