Pin dependencies; update to latest versions.
Closes https://github.com/MerosCrypto/Meros/issues/297.

Nim 1.2.12 was upgraded to due to Nimble < 0.13 being insecure. Nim 
1.2.10 and 1.2.12 are the 1.2 releases which ship with 0.13, 1.2.12 
being preferred as it has more bug fixes.

Notably pins to chronos v2 when v3 recently came out. This is due to 
notable changes with exception handling in v3; the same reason we don't 
use Nim 1.4.

Uses a dated copy of nim-chronicles to maintain the existing node color 
scheme. The new releases don't add anything meaningful (a minor bug fix 
on Windows + a basic code execution timer).

Removes the BLS code which reduces scalars as that is now part of mc_bls 
directly.