Remove the ed25519 library for the one present in-tree.
Moves ed25519 files to under Ristretto. Part of preparing to use 
Ristretto > Ed25519.

Regenerates vectors as a proof the key handling is the same. That said, 
several vector files have mutations. This is actually due to the RPC 
redo, and now some vectors no longer have the `argon` field which is no 
longer serialized into JSON. The regenerated vectors have still been 
merged to demonstrate the keys/signatures are untouched. Each file was 
manually checked to ensure that was the only difference.