Mild lint.
The plan was to handle UnusedImports AND ObservableStores. That said, I 
couldn't disable the latter using push or pragma blocks. Their actual 
suggested resolution also didn't work, notably in SyncManagerObj and so 
on when the variables were ONLY used inside that function. Because of 
that, this just does some mild variable moves that are beneficial 
anyways.