Implement a getTransactionHistory function as needed for the GUI.
This should NOT be used by services as it's currently inaccurate. While 
it includes payments in with the same threshold as getUTXOs, it has a 
much looser stance on payments out (where it'll naively select a 'best 
option' from unconfirmed options). While this works for the GUI, it is 
not up to spec for what services need, which is clearly documented in 
its definition.

On a branch as it's untested and needed by the new GUI (not a blocker 
but I would like to wait and see it used in the new GUI). Has 
implications for https://github.com/MerosCrypto/Meros/issues/292.