Tweaks and fixes.
Updates cache initialization to take in the result from Epochs instead 
of independently calculating it (which is no longer feasible).

Changes the one remaining instance of height - 5 to height - 10. 
TransactionStatus.epoch is still corrupted. We can safely set it to 
first mentioned + 11, which will aggresively kill Merit yet support 
bring ups. Right now, it's set to + 6.