Move to Bech32m.
Grabs a pretty new Python package to do so (10 days) since the existing 
one was never updated. I have concerns over that and it may be better 
simply to ship it ourselves, as it's just a package of sipa's work. That 
said, it's a direct package of sipa's work which I've verified, so it 
should be good.

RIP the Address test which malleated Bech32 as a test case.